Send notifications on save
By default we now don't send any notifications when saving (these were
being sent due to a bug in Gravity Forms 1.8, see issue #12). A new
setting has been added to the notifications screen where a notification
can be forced to be sent if the "Send on save" setting is checked for
the notification.